backend/app/services/rag/vector_store.py
import faiss
import numpy as np


def create_faiss_index(embeddings: np.ndarray):
    """
    Create a FAISS index from embeddings.
    """
    dimension = embeddings.shape[1]
    index = faiss.IndexFlatL2(dimension)
    index.add(embeddings)
    return index